Redefine the VM primary node assumptions
Previously, we used the NUMA node where most of the VM memory was
allocated to make the VM "belong" to that node. Now we have 2 types of
primary nodes: the memory and vcpu one. This tool mostly focuses on cpu
usage, so a VM now belongs to the node where all of its vcpus are
running. If the VM is floating between nodes or has a vcpus running on
different nodes, we warn and fallback to using the memory primary node.

Signed-off-by: Julien Desfossez <jdesfossez@digitalocean.com>